Streamline GLibPointerArray handling
