+++ src/Generation/Generator/Renderer/Internal/ReturnType/Converter/GLibPointerArrayCallback.cs
@@ -1,10 +1,10 @@
 namespace Generator.Renderer.Internal.ReturnType;
 
-internal class InterfaceGLibPtrArray : ReturnTypeConverter
+internal class GLibPointerArrayCallback : ReturnTypeConverter
 {
     public bool Supports(GirModel.ReturnType returnType)
     {
-        return returnType.AnyType.IsGLibPtrArray<GirModel.Interface>();
+        return returnType.AnyType.IsGLibPtrArray();
     }
 
     public RenderableReturnType Convert(GirModel.ReturnType returnType)
